# Further parser optimizations