See https://github.com/sebastienros/esprima-dotnet/pull/314#issuecomment-1219433726

> I've found some low hanging fruits around Scanner.ScanPunctuator:


## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|21.195 ms|0.9754 ms|1000.0000|468.7500|-|5,744 KB|
| **New** |	|	| **19.397 ms (-8%)** | **1.4790 ms** | **781.2500 (-22%)** | **375.0000 (-20%)** | **-** | **4,743 KB (-17%)** |
| Old |ParseProgram|backbone-1.1.0|2.589 ms|0.1904 ms|156.2500|74.2188|-|888 KB|
| **New** |	|	| **2.678 ms (+3%)** | **0.0325 ms** | **144.5313 (-7%)** | **66.4063 (-11%)** | **-** | **727 KB (-18%)** |
| Old |ParseProgram|jquery-1.9.1|15.906 ms|1.5005 ms|843.7500|406.2500|-|4,991 KB|
| **New** |	|	| **15.599 ms (-2%)** | **2.0704 ms** | **656.2500 (-22%)** | **328.1250 (-19%)** | **-** | **4,106 KB (-18%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|27.866 ms|25.7253 ms|1312.5000|656.2500|187.5000|7,678 KB|
| **New** |	|	| **23.914 ms (-14%)** | **0.8037 ms** | **1031.2500 (-21%)** | **500.0000 (-24%)** | **-** | **6,346 KB (-17%)** |
| Old |ParseProgram|mootools-1.4.5|12.806 ms|1.4193 ms|671.8750|328.1250|-|4,028 KB|
| **New** |	|	| **12.576 ms (-2%)** | **3.1946 ms** | **578.1250 (-14%)** | **250.0000 (-24%)** | **-** | **3,315 KB (-18%)** |
| Old |ParseProgram|underscore-1.5.2|2.247 ms|0.0597 ms|136.7188|58.5938|-|748 KB|
| **New** |	|	| **2.279 ms (+1%)** | **0.1745 ms** | **109.3750 (-20%)** | **54.6875 (-7%)** | **-** | **607 KB (-19%)** |
| Old |ParseProgram|yui-3.12.0|11.626 ms|2.0172 ms|656.2500|281.2500|-|3,789 KB|
| **New** |	|	| **11.477 ms (-1%)** | **4.0584 ms** | **500.0000 (-24%)** | **250.0000 (-11%)** | **-** | **3,096 KB (-18%)** |