@CharlieEriksen 2022-08-19
I realize this is closed, but this is some brilliant work! Major kudos.